Prevent Input of Numbers That Exceed Data Type Limits #67
This commit resolves the issue where users were able to input numbers that exceeded the limits of the data types used for processing. Such input could lead to a java.lang.NumberFormatException when processing, causing errors in the application.

- Changes have been made to the 'dosages,' 'frequency,' and 'take' functionalities to prevent users from inputting numbers that are too large.
- Now, when adding a prescription, input validation checks ensure that dosages, frequency, and take values are within the allowed data type limits.

This update enhances the stability and reliability of the application by preventing potential NumberFormatException errors caused by oversized numbers in user input.

Fixes #67